Rename getThermocyclerProfileRepetitionsForPython.ts -> thermocyclerProfileConversions.ts.
