+++ step-generation/src/utils/index.ts
@@ -4,7 +4,6 @@ import { absorbanceReaderCollision } from './absorbanceReaderCollision'
 import { commandCreatorsTimeline } from './commandCreatorsTimeline'
 import { curryCommandCreator, curryWithoutPython } from './curryCommandCreator'
 import { getLabwareSlot } from './getLabwareSlot'
-import { getThermocyclerProfileRepetitionsForPython } from './getThermocyclerProfileRepetitionsForPython'
 import { modulePipetteCollision } from './modulePipetteCollision'
 import { reduceCommandCreators } from './reduceCommandCreators'
 import { thermocyclerPipetteCollision } from './thermocyclerPipetteCollision'
@@ -18,7 +17,6 @@ export {
   modulePipetteCollision,
   thermocyclerPipetteCollision,
   getLabwareSlot,
-  getThermocyclerProfileRepetitionsForPython,
 }
 export * from './commandCreatorArgsGetters'
 export * from './constructInvariantContextFromAnalysis'
@@ -30,4 +28,6 @@ export * from './misc'
 export * from './pythonFileUtils'
 export * from './pythonFormat'
 export * from './safePipetteMovements'
+export * from './thermocyclerProfileConversions'
+
 export const uuid: () => string = uuidv4
